Move shell completion functions into completion.go
